cleaning up static items
